Move question exceptions
