+++ edsl/data/exceptions.py
@@ -1,5 +1,4 @@
-from edsl.exceptions.BaseException import BaseException
-
+from ..base import BaseException
 
 class CacheError(BaseException):
     relevant_doc = "https://docs.expectedparrot.com/en/latest/agents.html#agent-lists"
