+++ edsl/questions/answer_validator_mixin.py
@@ -2,7 +2,7 @@
 
 import re
 from typing import Any, Type, Union
-from edsl.exceptions.questions import (
+from .exceptions import (
     QuestionAnswerValidationError,
 )
