+++ edsl/questions/question_free_text.py
@@ -6,9 +6,9 @@ from uuid import uuid4
 
 from pydantic import field_validator, model_validator, BaseModel
 
-from ..exceptions.questions import QuestionAnswerValidationError
 from ..prompts import Prompt
 
+from .exceptions import QuestionAnswerValidationError
 from .question_base import QuestionBase
 from .response_validator_abc import ResponseValidatorABC
 from .decorators import inject_exception
