+++ tests/questions/test_QuestionRank.py
@@ -1,5 +1,5 @@
 import pytest
-from edsl.exceptions.questions import (
+from edsl.questions.exceptions import (
     QuestionAnswerValidationError,
     QuestionResponseValidationError,
 )
