+++ tests/questions/test_ResponseValidator.py
@@ -3,7 +3,7 @@ from pydantic import BaseModel, Field
 from typing import Optional, Any
 from decimal import Decimal
 
-from edsl.exceptions.questions import QuestionAnswerValidationError
+from edsl.questions.exceptions import QuestionAnswerValidationError
 from abc import ABC, abstractmethod
 
 from edsl.questions.response_validator_abc import ResponseValidatorABC
